+++ zebra/zebra_rnh.h
@@ -29,40 +29,6 @@
 extern "C" {
 #endif
 
-typedef enum { RNH_NEXTHOP_TYPE, RNH_IMPORT_CHECK_TYPE } rnh_type_t;
-
-/* Nexthop structure. */
-struct rnh {
-	uint8_t flags;
-
-#define ZEBRA_NHT_CONNECTED  	0x1
-#define ZEBRA_NHT_DELETED       0x2
-#define ZEBRA_NHT_EXACT_MATCH   0x4
-
-	/* VRF identifier. */
-	vrf_id_t vrf_id;
-
-	afi_t afi;
-
-	rnh_type_t type;
-
-	uint32_t seqno;
-
-	struct route_entry *state;
-	struct prefix resolved_route;
-	struct list *client_list;
-
-	/* pseudowires dependent on this nh */
-	struct list *zebra_pseudowire_list;
-
-	struct route_node *node;
-
-	/*
-	 * if this has been filtered for the client
-	 */
-	int filtered[ZEBRA_ROUTE_MAX];
-};
-
 extern int zebra_rnh_ip_default_route;
 extern int zebra_rnh_ipv6_default_route;
